refactor: 매장 발주 스토어 물리 분리 및 매장 화면 공통 UI 유틸 추출
