+++ src/views/store/inbound/StoreInboundListView.vue
@@ -5,6 +5,7 @@ import AppLayout from '@/components/common/AppLayout.vue'
 import { roleMenus } from '@/config/roleMenus.js'
 import { useAuthStore } from '@/stores/auth.js'
 import { useStoreInboundStore } from '@/stores/storeInbound.js'
+import { buildHeadline, formatDate, formatDateTime, storeInboundStatusClass } from '@/features/store/common/ui.js'
 
 const router = useRouter()
 const auth = useAuthStore()
@@ -38,36 +39,12 @@ const arrivedOrders = computed(() =>
 )
 const arrivedOrderCount = computed(() => arrivedOrders.value.length)
 
-function formatDateTime(iso) {
-  if (!iso) return '-'
-  const date = new Date(iso)
-  const pad = (value) => String(value).padStart(2, '0')
-  return `${date.getFullYear()}-${pad(date.getMonth() + 1)}-${pad(date.getDate())} ${pad(date.getHours())}:${pad(date.getMinutes())}`
-}
-
-function formatDate(iso) {
-  if (!iso) return '-'
-  const date = new Date(iso)
-  const pad = (value) => String(value).padStart(2, '0')
-  return `${date.getFullYear()}-${pad(date.getMonth() + 1)}-${pad(date.getDate())}`
-}
-
 function headlineLabel(order) {
-  if (!order || order.items.length === 0) return '-'
-  return order.items.length > 1
-    ? `${order.items[0].productName} 외 ${order.items.length - 1}건`
-    : order.items[0].productName
+  return buildHeadline(order?.items)
 }
 
 function inboundStatusClass(status) {
-  return (
-    {
-      READY_TO_SHIP: 'bg-slate-100 text-slate-700',
-      IN_TRANSIT: 'bg-blue-100 text-blue-700',
-      ARRIVED: 'bg-amber-100 text-amber-700',
-      RECEIVED: 'bg-[#EBF5F5] text-black',
-    }[status] ?? 'bg-gray-100 text-gray-600'
-  )
+  return storeInboundStatusClass(status)
 }
 
 function inboundActionLabel(order) {
